Move function merge two projects from Mastodon plugins menu to Fiji plugin menu
* Reasoning: this method does not really touch the currently open Mastodon session but merges two projects outside of it
